Create EVs_and_Repel_rebalance.rb
EV mod and repel mod.

What this does is make the genuis wings give 252 EVs.

The reason I chose genius wings and not vitmains is because they require you to get to the point in the game where EVs start becoming important. 

The repel mod is a rebalancing to make each of them cost about the same amount for the number of steps you get. In addition, this can be changed. The mod has options within the file that you can change with a text editor.